+++ src/main/kotlin/me/beresnev/kdiameter/network/message/DiameterMessageDecoder.kt
@@ -5,10 +5,12 @@ import me.beresnev.kdiameter.extensions.readFourBytes
 import me.beresnev.kdiameter.extensions.readThreeBytes
 import me.beresnev.kdiameter.extensions.toUnsignedLong
 import me.beresnev.kdiameter.model.Avp
+import me.beresnev.kdiameter.network.message.flags.AvpFlags
+import me.beresnev.kdiameter.network.message.flags.CommandFlags
 import java.io.ByteArrayInputStream
 import java.io.InputStream
 
-class DiameterMessageDecoder {
+object DiameterMessageDecoder {
 
     /** 0               1               2               3
      *  0 1 2 3 4 5 6 7 0 1 2 3 4 5 6 7 0 1 2 3 4 5 6 7 0 1 2 3 4 5 6 7
@@ -29,11 +31,11 @@ class DiameterMessageDecoder {
     fun decode(message: ByteArray): DiameterMessage {
         val dataStream = ByteArrayInputStream(message)
 
-        assertMessageVersionIsSupported(dataStream.readByte()) // first byte is version
-        assertMessageLength(message.size, dataStream) // next three are for message length
+        assertMessageVersionIsSupported(dataStream.readByte())
+        assertMessageLength(message.size, dataStream)
 
-        val commandFlags = CommandFlags(dataStream.readByte()) // next byte is for command flags
-        val commandCode = dataStream.readThreeBytes() // next three are for command code
+        val commandFlags = CommandFlags(dataStream.readByte())
+        val commandCode = dataStream.readThreeBytes()
 
         // while in RFC it says nothing of type (signed/unsigned), and since
         // applicationIds are user defined, we assume it can be unsigned
@@ -71,7 +73,78 @@ class DiameterMessageDecoder {
         }
     }
 
+    fun decodeAvps(data: ByteArray): List<Avp> {
+        return decodeAvps(ByteArrayInputStream(data))
+    }
+
     private fun decodeAvps(dataStream: InputStream): List<Avp> {
-        return emptyList()
+        if (dataStream.available() == 0) {
+            return emptyList()
+        }
+
+        val avps = ArrayList<Avp>()
+        while (dataStream.available() > 0) {
+            avps.add(decodeAvp(dataStream))
+        }
+        return avps
+    }
+
+    /**
+     * 0               1               2               3
+     * 0 1 2 3 4 5 6 7 0 1 2 3 4 5 6 7 0 1 2 3 4 5 6 7 0 1 2 3 4 5 6 7
+     * +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+     * |                           AVP Code                          |
+     * +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+     * | V M P r r r r r|                  AVP Length                |
+     * +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+     * |                        Vendor-ID (opt)                      |
+     * +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+     * |                            Data ...                         |
+     * +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
+     */
+    private fun decodeAvp(dataStream: InputStream): Avp {
+        val avpCode = dataStream.readFourBytes().toUnsignedLong()
+        val avpFlags = AvpFlags(dataStream.readByte())
+
+        // The AVP Length field is three octets, and indicates the number of
+        // octets in this AVP including the AVP Code field, AVP Length field,
+        // AVP Flags field, Vendor-ID field (if present), and the AVP Data
+        // field.  If a message is received with an invalid attribute length,
+        // the message MUST be rejected.
+        val avpLength = dataStream.readThreeBytes()
+
+        val hasVendorId = avpFlags.isVendorSpecific
+        val vendorId = if (hasVendorId) dataStream.readFourBytes().toUnsignedLong() else null
+
+        // since we already read some of the bytes from the stream,
+        // we need to calculate how many bytes are left for data
+        // dataLength = avpLength - 4(code) - 1(flags) - 3(length) [- 4 (vendor)]
+        val dataLength = avpLength - 8 - if (hasVendorId) 4 else 0
+
+        val rawData = ByteArray(dataLength)
+        dataStream.read(rawData, 0, dataLength)
+
+        if (avpLength % 4 != 0) { // value not aligned, must skip empty padding bytes
+            skipPadding(avpLength, dataStream)
+        }
+        return Avp.create(avpCode, avpFlags, vendorId, rawData)
+    }
+
+    /**
+     * Each AVP of type OctetString MUST be padded to align on a 32-bit
+     * boundary, while other AVP types align naturally.  A number of zero-
+     * valued bytes are added to the end of the AVP Data field until a word
+     * boundary is reached.
+     *
+     * NOTE! "The length of the padding is not reflected in the AVP Length field."
+     * This is the reason we need to skip empty padding bytes.
+     */
+    private fun skipPadding(avpLength: Int, dataStream: InputStream) {
+        var avpLengthWithPadding = avpLength.toLong()
+        do {
+            // https://en.wikipedia.org/wiki/Data_structure_alignment#Computing_padding
+            val padding = (4L - avpLengthWithPadding % 4L) % 4L
+            avpLengthWithPadding += dataStream.skip(padding)
+        } while (avpLengthWithPadding % 4L != 0L)
     }
 }